LayoutTests/media/video-interruption-with-resume-not-allowing-play-expected.txt


Test that play() during interruption does nothing, ending interruption does not allow playback to resume.

EVENT(canplaythrough)

RUN(internals.beginMediaSessionInterruption())
RUN(video.play())
100ms timer fired...
EXPECTED (video.paused == 'true') OK
RUN(internals.endMediaSessionInterruption(''))
100ms timer fired...
EXPECTED (video.paused == 'true') OK

END OF TEST
